backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyMuPDF==1.24.9
pdfplumber==0.11.4
keybert==0.8.5
sentence-transformers==3.0.1
scikit-learn==1.5.1
numpy==2.0.1
requests==2.32.3
